SNR colours, SNR labels for links

+++ app/src/main/java/com/atakmap/android/soothsayer/models/linksmodel/LinkResponse.kt
@@ -44,7 +44,7 @@ data class Receiver(
     @SerializedName("Receiver gain dBd")
     val receiverGainDBd: Double,
     @SerializedName("Receiver gain dBi")
-    val receiverGainDBi: Int
+    val receiverGainDBi: Double
 )
 
 data class Transmitter(
